+++ lib/apex_charts/renderer.rb
@@ -6,36 +6,39 @@ require_relative 'version'
 module ApexCharts
   class Renderer
     class << self
-      def render_default(options)
-        renderer = new options
-
-        html = if renderer.id_number == '1' && !ApexCharts.config.default_options.empty?
-                 renderer.window_apex
-               else
-                 ''
-               end
-        html_chart_rendering = <<~createJSChart
-        // ApexCharts.RB #{RELEASE}
-        var #{renderer.variable} = new ApexCharts(document.querySelector("##{renderer.element_id}"), #{substitute_function_object(renderer.options.to_json)});
-        #{renderer.variable}.render();
-        createJSChart
-
-        div = <<~JS
-          <div id="#{renderer.element_id}" class="#{renderer.css_class}" style="#{renderer.style}"></div>
-        JS
-        js = if renderer.options[:defer]
-               deferred_js(html_chart_rendering)
-             else
-               non_deferred_js(html_chart_rendering)
-             end
-        html + div + js
+      def render(options)
+        new(options).render
       end
+    end
 
-      def deferred_js(html_chart_rendering)
-        js = <<~JS
-        <script type="text/javascript">
+    attr_reader :options
+
+    def initialize(options)
+      @options = options
+    end
+
+    def render
+      html = ''
+      html = window_apex if id_number == '1' && !ApexCharts.config.default_options.empty?
+
+      chart_rendering = <<~JS
+        var #{variable} = new ApexCharts(document.querySelector("##{element_id}"), #{substitute_function_object(options.to_json)});
+        #{variable}.render();
+      JS
+
+      html += <<~HTML
+        <div id="#{element_id}" class="#{css_class}" style="#{style}"></div>
+        #{script(defer(chart_rendering))}
+      HTML
+    end
+
+    def defer(js)
+      if defer?
+        <<~DEFERRED
           (function() {
-            var createChart = function() { #{html_chart_rendering} };
+            var createChart = function() {
+              #{indent(js)}
+            };
             if (window.addEventListener) {
               window.addEventListener("load", createChart, true);
             } else if (window.attachEvent) {
@@ -44,30 +47,21 @@ module ApexCharts
               createChart();
             }
           })();
-        </script>
-        JS
-      end
-
-      def non_deferred_js(html_chart_rendering)
-        js = <<~JS
-        <script type="text/javascript">
-            #{html_chart_rendering}
-        </script>
-        JS
-      end
-
-      def substitute_function_object(json)
-        json.gsub(%r[{"function":{"args":"(?<args>.*?)","body":"(?<body>.*?)"}}]) do
-          body = "\"#{$~&.[](:body)}\"".undump
-          "function(#{$~&.[](:args)}){#{body}}"
-        end
+        DEFERRED
+      else
+        js
       end
     end
 
-    attr_reader :options
+    def substitute_function_object(json)
+      json.gsub(%r[{"function":{"args":"(?<args>.*?)","body":"(?<body>.*?)"}}]) do
+        body = "\"#{$~&.[](:body)}\"".undump
+        "function(#{$~&.[](:args)}){#{body}}"
+      end
+    end
 
-    def initialize(options)
-      @options = options
+    def defer?
+      @defer ||= options.delete(:defer)
     end
 
     def attributes
@@ -99,11 +93,21 @@ module ApexCharts
     end
 
     def window_apex
-      <<~HTML
-        <script type="text/javascript">
-          window.Apex = #{ApexCharts.config.default_options.to_json}
+      script("window.Apex = #{ApexCharts.config.default_options.to_json}")
+    end
+
+    def script(js)
+      <<~SCRIPT
+        <script type="text/javascript" apexcharts-rb="#{RELEASE}">
+        #{js}
         </script>
-      HTML
+      SCRIPT
+    end
+
+    def indent(content, times=2)
+      content.lines.map.with_index do |line, index|
+        (index == 0 ? '' : '  ' * times) + line
+      end.join
     end
   end
 end